backend/requirements.txt
flask
flask-cors
opencv-python
onnxruntime
numpy
python-dotenv
ultralytics
pytesseract
Pillow
tensorflow
openai
